--- backend/src/main/java/com/bungeobbang/backend/member/domain/OauthUserInfo.java
@@ -1,5 +1,0 @@
-package com.bungeobbang.backend.member.domain;
-
-public interface OauthUserInfo {
-    String getSocialLoginId();
-}
