+++ backend/src/main/java/com/bungeobbang/backend/member/infrastructure/oauthuserinfo/OauthUserInfo.java
@@ -1,0 +1,5 @@
+package com.bungeobbang.backend.member.infrastructure.oauthuserinfo;
+
+public interface OauthUserInfo {
+    String getSocialLoginId();
+}
